move useEffect inside of Rockets.jsx
